修改  short -> int

+++ src/main/java/cn/t/tool/netproxytool/http/constants/HttpProxyServerConfig.java
@@ -10,7 +10,7 @@ import io.netty.handler.logging.LogLevel;
  * @since 2020-02-22 23:00
  **/
 public class HttpProxyServerConfig {
-    public static final short SERVER_PORT = 1087;
+    public static final int SERVER_PORT = 1087;
 
     public static final int HTTP_PROXY_READ_TIME_OUT_IN_SECONDS = 0;
     public static final int HTTP_PROXY_WRITE_TIME_OUT_IN_SECONDS = 0;
